Update array_sort to use Shellsort instead of Heapsort
This particular version of Shellsort is faster, simpler and uses less memory than the Heapsort version previously included in the library.

I was looking for another non-recursive strictly-in-place sorting algorithm because Heapsort usually hits a “cache miss” wall when sorting large (~10^7) arrays.

This particular version of Shellsort (with a growing factor of 2.25 rather than the popular 3) checks all the boxes and performs very well in practice. Give it a try before dismissing it!